fabric_get: Refactor router
# Summary

Refactor  the router for the following endpoint:

Verb: GET
Path: `/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}`

## Move endpoint handler to a new file in a directory aligned with NDFC endpoint path

v1/endpoints/lan_fabric/rest/config/fabrics/fabric_get.py

## ./app/main.py

- Update imports
- Add app.include_router(fabric_get.router, tags=["Fabrics (v1)"])
- Modify all other tags to include the API version.

## ./app/v1/endpoints/fabric.py

- v1_get_fabric_by_fabric_name() - Remove old fabric_get handler
