+++ app/v1/endpoints/fabric.py
@@ -34,23 +34,6 @@ def build_nv_pairs(fabric):
     return fabric.model_dump()
 
 
-@app.get(
-    "/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}",
-    response_model=FabricResponseModel,
-)
-def v1_get_fabric_by_fabric_name(*, session: Session = Depends(get_session), fabric_name: str):
-    """
-    # Summary
-
-    GET request handler with fabric_name as path parameter.
-    """
-    db_fabric = session.exec(select(Fabric).where(Fabric.FABRIC_NAME == fabric_name)).first()
-    if not db_fabric:
-        raise HTTPException(status_code=404, detail=f"Fabric {fabric_name} not found")
-    response = build_response(db_fabric)
-    return response
-
-
 @app.post(
     "/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/{template_name}",
     response_model=FabricResponseModel,
